+++ spectral_cube/spectral_cube.py
@@ -24,6 +24,8 @@ from astropy import stats
 
 import numpy as np
 
+from radio_beam import Beam
+
 from . import cube_utils
 from . import wcs_utils
 from . import spectral_axis
@@ -186,14 +188,6 @@ class BaseSpectralCube(BaseNDClass, MaskableArrayMixinClass,
         cube._spectral_scale = spectral_axis.wcs_unit_scale(spectral_unit)
 
         return cube
-
-    def _attach_beam(self):
-
-        beam = cube_utils.try_load_beam(self.header)
-
-        if beam is not None:
-            self.beam = beam
-            self._meta['beam'] = beam
 
     @property
     def unit(self):
@@ -2275,7 +2269,7 @@ class SpectralCube(BaseSpectralCube):
 
     def __init__(self, data, wcs, mask=None, meta=None, fill_value=np.nan,
                  header=None, allow_huge_operations=False, beam=None,
-                 read_beam=True, wcs_tolerance=0.0, **kwargs):
+                 wcs_tolerance=0.0, **kwargs):
 
         super(SpectralCube, self).__init__(data=data, wcs=wcs, mask=mask,
                                            meta=meta, fill_value=fill_value,
@@ -2285,36 +2279,18 @@ class SpectralCube(BaseSpectralCube):
                                            **kwargs)
 
         # Beam loading must happen *after* WCS is read
-        if beam is None and read_beam:
-            self._attach_beam()
 
-        if beam is None and not read_beam and 'BUNIT' in self._meta:
-            bunit = re.sub("\s", "", self._meta['BUNIT'].lower())
-            if bunit == 'jy/beam':
-                warnings.warn("Units are in Jy/beam. Attempting to parse "
-                              "header for beam information.")
-
-                self._attach_beam()
-
-                if hasattr(self, 'beam') or hasattr(self, 'beams'):
-                    warnings.warn("Units were Jy/beam.  The 'beam' is now "
-                                  "stored in the .beam attribute, and the "
-                                  "units are set to Jy")
-                else:
-                    warnings.warn("Could not parse Jy/beam unit.  Either "
-                                  "you should install the radio_beam "
-                                  "package or manually replace the units."
-                                  " For now, the units are being interpreted "
-                                  "as Jy.")
+        if beam is None:
+            beam = cube_utils.try_load_beam(self.header)
+        else:
+            if not isinstance(beam, Beam):
+                raise TypeError("beam must be a radio_beam.Beam object.")
 
         if beam is not None:
             self.beam = beam
             self._meta['beam'] = beam
-
-            # Ensure that the beam is properly defined in the header
             self._header.update(beam.to_header_keywords())
 
-        if 'beam' in self._meta:
             self.pixels_per_beam = (self.beam.sr /
                                     (astropy.wcs.utils.proj_plane_pixel_area(self.wcs) *
                                      u.deg**2)).to(u.dimensionless_unscaled).value
@@ -2327,6 +2303,30 @@ class SpectralCube(BaseSpectralCube):
         return newcube
 
     _new_cube_with.__doc__ = BaseSpectralCube._new_cube_with.__doc__
+
+    def with_beam(self, beam):
+        '''
+        Attach a beam object to the `~SpectralCube`.
+
+        Parameters
+        ----------
+        beam : `~radio_beam.Beam`
+            `Beam` object defining the resolution element of the
+            `~SpectralCube`.
+        '''
+
+        if not isinstance(beam, Beam):
+            raise TypeError("beam must be a radio_beam.Beam object.")
+
+        meta = self._meta.copy()
+        meta['beam'] = beam
+
+        header = self._header.copy()
+        header.update(beam.to_header_keywords())
+
+        newcube = self._new_cube_with(meta=self.meta, beam=beam)
+
+        return newcube
 
     def spatial_smooth_median(self, ksize, **kwargs):
         """
@@ -2699,8 +2699,7 @@ class SpectralCube(BaseSpectralCube):
                                        normalize_kernel=True)
             pb.update()
 
-        newcube = self._new_cube_with(data=newdata, read_beam=False,
-                                      beam=beam)
+        newcube = self._new_cube_with(data=newdata, beam=beam)
 
         return newcube
 
@@ -2740,7 +2739,6 @@ class VaryingResolutionSpectralCube(BaseSpectralCube):
             'pa':5.0}``
         """
         # these types of cube are undefined without the radio_beam package
-        from radio_beam import Beam
 
         beam_table = kwargs.pop('beam_table', None)
         beams = kwargs.pop('beams', None)
@@ -2853,6 +2851,16 @@ class VaryingResolutionSpectralCube(BaseSpectralCube):
 
             # Slice objects know how to parse Beam objects stored in the
             # metadata
+            # A 2D slice with a VRSC should not be allowed along a
+            # position-spectral axis
+            if not isinstance(self.beams[specslice], Beam):
+                raise AttributeError("2D slices along a spectral axis are not "
+                                     "allowed for "
+                                     "VaryingResolutionSpectralCubes. Convolve"
+                                     " to a common resolution with "
+                                     "`convolve_to` before attempting "
+                                     "position-spectral slicing.")
+
             meta['beam'] = self.beams[specslice]
             return Slice(value=self.filled_data[view],
                          wcs=newwcs,
@@ -2995,7 +3003,6 @@ class VaryingResolutionSpectralCube(BaseSpectralCube):
         includemask : np.array
             A boolean array where ``True`` indicates the good beams
         """
-        from radio_beam import Beam
 
         includemask = np.ones(len(self.beams), dtype='bool')
 
@@ -3233,7 +3240,6 @@ class VaryingResolutionSpectralCube(BaseSpectralCube):
                                meta=self.meta, fill_value=self.fill_value,
                                header=self.header,
                                allow_huge_operations=self.allow_huge_operations,
-                               read_beam=False,
                                beam=beam,
                                wcs_tolerance=self._wcs_tolerance)